refactor: update demo token generation to use string IDs
- Changed demo user and organization ID retrieval in AuthService to use string values instead of converting to integers.
- This change simplifies the token generation process and aligns with the updated configuration structure.

+++ backend/src/main/kotlin/com/moneat/services/AuthService.kt
@@ -373,8 +373,8 @@ class AuthService {
     }
 
     fun generateDemoToken(): String {
-        val demoUserId = com.moneat.config.EnvConfig.Demo.USER_ID.toInt()
-        val demoOrgId = com.moneat.config.EnvConfig.Demo.ORG_ID.toInt()
+        val demoUserId = com.moneat.config.EnvConfig.Demo.USER_ID
+        val demoOrgId = com.moneat.config.EnvConfig.Demo.ORG_ID
         val demoEpochMs = com.moneat.config.EnvConfig.Demo.epochMs
 
         return JWT.create()
